src/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
matplotlib
seaborn
geopandas
folium
requests
google-generativeai
